+++ Extra/XCode Templates/VCore (SwiftUI)/View (Dynamic UI Model).xctemplate/___FILEBASENAME___Parameters.swift
@@ -2,8 +2,8 @@
 
 import Foundation
 
-// MARK: - ___VARIABLE_productName___ View Parameters
-struct ___VARIABLE_productName___ViewParameters {
+// MARK: - ___VARIABLE_productName___ Parameters
+struct ___VARIABLE_productName___Parameters {
     // MARK: Properties
     
     // MARK: Mock