+++ Extra/XCode Templates/VCore (UIKit)/UITableViewCell.xctemplate/___FILEBASENAME___.swift
@@ -3,22 +3,17 @@
 import UIKit
 import VCore
 
-// MARK: - ___VARIABLE_productName___ View
-final class ___VARIABLE_productName___View: UIView {
+// MARK: - ___VARIABLE_productName___
+final class ___VARIABLE_productName___: UITableViewCell, ConfigurableUITableViewCell {
     // MARK: Subviews
     
     // MARK: Properties
-    private typealias UIModel = ___VARIABLE_productName___ViewUIModel
+    private typealias UIModel = ___VARIABLE_productName___UIModel
     
     // MARK: Initializers
-    init() {
-        super.init(frame: .zero)
+    override init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
+        super.init(style: style, reuseIdentifier: reuseIdentifier)
         setUp()
-    }
-    
-    convenience init(parameters: ___VARIABLE_productName___ViewParameters) {
-        self.init()
-        configure(parameters: parameters)
     }
     
     required init?(coder: NSCoder) {
@@ -34,6 +29,7 @@ final class ___VARIABLE_productName___View: UIView {
     
     private func setUpView() {
         backgroundColor = UIModel.Colors.background
+        selectionStyle = .none
     }
     
     private func addSubviews() {
@@ -46,8 +42,8 @@ final class ___VARIABLE_productName___View: UIView {
         ])
     }
     
-    // MARK: Configuration
-    func configure(parameters: ___VARIABLE_productName___ViewParameters) {
-        
+    // MARK: Configurable Table View Cell
+    func configure(parameter: some UITableViewCellParameter) {
+        guard let parameters = parameter as? ___VARIABLE_productName___Parameters else { return }
     }
 }