+++ Extra/XCode Templates/VCore (UIKit)/UITableViewCell.xctemplate/___FILEBASENAME___Parameters.swift
@@ -1,0 +1,12 @@
+//  ___FILEHEADER___
+
+import Foundation
+import VCore
+
+// MARK: - ___VARIABLE_productName___ Parameters
+struct ___VARIABLE_productName___Parameters: UITableViewCellParameter {
+    // MARK: Properties
+    
+    // MARK: Table View Cell Parameter
+    var reuseID: String { ___VARIABLE_productName___.reuseID }
+}