+++ Extra/XCode Templates/VCore (UIKit)/UIView (Dynamic UI Model).xctemplate/___FILEBASENAME___.swift
@@ -3,16 +3,16 @@
 import UIKit
 import VCore
 
-// MARK: - ___VARIABLE_productName___ View
-final class ___VARIABLE_productName___View: UIView {
+// MARK: - ___VARIABLE_productName___
+final class ___VARIABLE_productName____: UIView {
     // MARK: Subviews
     
     // MARK: Properties
-    private var uiModel: ___VARIABLE_productName___ViewUIModel
+    private var uiModel: ___VARIABLE_productName____UIModel
     
     // MARK: Initializers
     init(
-        uiModel: ___VARIABLE_productName___ViewUIModel = .init()
+        uiModel: ___VARIABLE_productName____UIModel = .init()
     ) {
         self.uiModel = uiModel
         super.init(frame: .zero)
@@ -21,8 +21,8 @@ final class ___VARIABLE_productName___View: UIView {
     }
     
     convenience init(
-        uiModel: ___VARIABLE_productName___ViewUIModel = .init(),
-        parameters: ___VARIABLE_productName___ViewParameters
+        uiModel: ___VARIABLE_productName____UIModel = .init(),
+        parameters: ___VARIABLE_productName____Parameters
     ) {
         self.init(uiModel: uiModel)
         configure(parameters: parameters)
@@ -54,12 +54,12 @@ final class ___VARIABLE_productName___View: UIView {
     }
     
     // MARK: Configuration - UI Model
-    func configure(uiModel: ___VARIABLE_productName___ViewUIModel) {
+    func configure(uiModel: ___VARIABLE_productName____UIModel) {
         self.uiModel = uiModel
     }
     
     // MARK: Configuration - Parameters
-    func configure(parameters: ___VARIABLE_productName___ViewParameters) {
+    func configure(parameters: ___VARIABLE_productName____Parameters) {
         
     }
 }