+++ Extra/XCode Templates/VCore (UIKit)/UIView + Dynamic UI Model.xctemplate/___FILEBASENAME___ViewParameters.swift
@@ -1,8 +1,0 @@
-//  ___FILEHEADER___
-
-import Foundation
-
-// MARK: - ___VARIABLE_productName___ View Parameters
-struct ___VARIABLE_productName___ViewParameters {
-    
-}